[Fix bug] Refactor Size API routes
