+++ app/api/[storeId]/sizes/route.ts
@@ -1,30 +1,10 @@
-import prismadb from "@/lib/prismadb";
 import { auth } from "@clerk/nextjs";
+import prismadb from "@/lib/prismadb";
 import { NextResponse } from "next/server";
 
-export async function GET(
-  _req: Request,
-  { params }: { params: { sizeId: string } }
-) {
-  try {
-    if (!params.sizeId) {
-      return new NextResponse("Size id is required", { status: 400 });
-    }
-
-    const size = await prismadb.size.findUnique({
-      where: { id: params.sizeId },
-    });
-
-    return NextResponse.json(size);
-  } catch (error) {
-    console.log("SIZE_GET", error);
-    return new NextResponse("Internal error", { status: 500 });
-  }
-}
-
-export async function PATCH(
+export async function POST(
   req: Request,
-  { params }: { params: { storeId: string; sizeId: string } }
+  { params }: { params: { storeId: string } }
 ) {
   try {
     const { userId } = auth();
@@ -43,8 +23,8 @@ export async function PATCH(
       return new NextResponse("Value is required", { status: 400 });
     }
 
-    if (!params.sizeId) {
-      return new NextResponse("Size id is required", { status: 400 });
+    if (!params.storeId) {
+      return new NextResponse("Store id is required", { status: 400 });
     }
 
     const storeByUserId = await prismadb.store.findFirst({
@@ -55,48 +35,33 @@ export async function PATCH(
       return new NextResponse("Unauthorized", { status: 403 });
     }
 
-    const size = await prismadb.size.updateMany({
-      where: { id: params.sizeId },
-      data: { name, value },
+    const size = await prismadb.size.create({
+      data: { name, value, storeId: params.storeId },
     });
 
     return NextResponse.json(size);
   } catch (error) {
-    console.log("SIZE_PATCH", error);
+    console.log("SIZES_POST", error);
     return new NextResponse("Internal error", { status: 500 });
   }
 }
 
-export async function DELETE(
+export async function GET(
   _req: Request,
-  { params }: { params: { storeId: string; sizeId: string } }
+  { params }: { params: { storeId: string } }
 ) {
   try {
-    const { userId } = auth();
-
-    if (!userId) {
-      return new NextResponse("Unauthenticated", { status: 401 });
+    if (!params.storeId) {
+      return new NextResponse("Store id is required", { status: 400 });
     }
 
-    if (!params.sizeId) {
-      return new NextResponse("Size id is required", { status: 400 });
-    }
-
-    const storeByUserId = await prismadb.store.findFirst({
-      where: { id: params.storeId, userId },
+    const sizes = await prismadb.size.findMany({
+      where: { storeId: params.storeId },
     });
 
-    if (!storeByUserId) {
-      return new NextResponse("Unauthorized", { status: 403 });
-    }
-
-    const size = await prismadb.size.deleteMany({
-      where: { id: params.sizeId },
-    });
-
-    return NextResponse.json(size);
+    return NextResponse.json(sizes);
   } catch (error) {
-    console.log("SIZE_DELETE", error);
+    console.log("SIZES_GET", error);
     return new NextResponse("Internal error", { status: 500 });
   }
 }
